backend/app/core/security.py
"""
»­åêÆ£¢#næüÆ£êÆ£â¸åüë
"""

from datetime import datetime, timedelta
from typing import Any, Optional, Union

from jose import jwt
from passlib.context import CryptContext
from fastapi import Depends, HTTPException, status
from fastapi.security import OAuth2PasswordBearer
from pydantic import ValidationError

from app.core.config import settings


# Ñ¹ïüÉÏÃ·å³óÆ­¹È
pwd_context = CryptContext(schemes=["bcrypt"], deprecated="auto")

# OAuth¹­üà
oauth2_scheme = OAuth2PasswordBearer(
    tokenUrl=f"{settings.API_V1_STR}/auth/login"
)


def verify_password(plain_password: str, hashed_password: str) -> bool:
    """s‡Ñ¹ïüÉhÏÃ·åÑ¹ïüÉ’Ô<Y‹"""
    return pwd_context.verify(plain_password, hashed_password)


def get_password_hash(password: str) -> str:
    """Ñ¹ïüÉ’ÏÃ·åY‹"""
    return pwd_context.hash(password)


def create_access_token(
    subject: Union[str, Any], expires_delta: Optional[timedelta] = None
) -> str:
    """¢¯»¹Èü¯ó’Y‹"""
    if expires_delta:
        expire = datetime.utcnow() + expires_delta
    else:
        expire = datetime.utcnow() + timedelta(
            minutes=settings.ACCESS_TOKEN_EXPIRE_MINUTES
        )
    
    to_encode = {"exp": expire, "sub": str(subject)}
    encoded_jwt = jwt.encode(
        to_encode, settings.SECRET_KEY, algorithm=settings.ALGORITHM
    )
    return encoded_jwt


def decode_access_token(token: str) -> dict:
    """¢¯»¹Èü¯ó’Ç³üÉY‹"""
    try:
        payload = jwt.decode(
            token, settings.SECRET_KEY, algorithms=[settings.ALGORITHM]
        )
        return payload
    except (jwt.JWTError, ValidationError) as e:
        raise HTTPException(
            status_code=status.HTTP_403_FORBIDDEN,
            detail="Could not validate credentials",
        )


# SupabasenJWTÈü¯ó’<Y‹¢p
def verify_supabase_token(token: str) -> dict:
    """SupabasenJWTÈü¯ó’<Y‹"""
    try:
        # Supabasen4oJWT_SECRET’(
        payload = jwt.decode(
            token, settings.SUPABASE_JWT_SECRET, algorithms=[settings.ALGORITHM]
        )
        return payload
    except (jwt.JWTError, ValidationError) as e:
        raise HTTPException(
            status_code=status.HTTP_403_FORBIDDEN,
            detail="Could not validate Supabase credentials",
        )
