backend/requirements.txt
# FastAPI関連
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic==2.4.2
pydantic-settings==2.0.3
email-validator==2.1.0.post1

# データベース
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
asyncpg==0.29.0

# データ処理
pandas==2.1.3
numpy==1.26.2
scipy==1.11.4
gpxpy==1.6.0
geopy==2.4.0

# セキュリティ
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1

# テスト
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.1

# Supabase
supabase==1.2.0

# ユーティリティ
python-multipart==0.0.6
aiofiles==23.2.1
